Properly distribute endpoints unit tests
